Fix printf format inconsistencies as pointed out by gcc -Wformat-signedness.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@15501

+++ coregrind/m_sigframe/sigframe-amd64-linux.c
@@ -494,7 +494,7 @@ void VG_(sigframe_create)( ThreadId tid,
    if (0)
       VG_(printf)("pushed signal frame; %%RSP now = %#lx, "
                   "next %%RIP = %#llx, status=%d\n",
-		  rsp, tst->arch.vex.guest_RIP, tst->status);
+		  rsp, tst->arch.vex.guest_RIP, (Int)tst->status);
 }
 
 
@@ -510,7 +510,7 @@ Bool restore_vg_sigframe ( ThreadState *tst,
 {
    if (frame->magicPI != 0x31415927 ||
        frame->magicE  != 0x27182818) {
-      VG_(message)(Vg_UserMsg, "Thread %d return signal frame "
+      VG_(message)(Vg_UserMsg, "Thread %u return signal frame "
                                "corrupted.  Killing process.\n",
 		   tst->tid);
       VG_(set_default_handler)(VKI_SIGSEGV);
@@ -594,7 +594,7 @@ void VG_(sigframe_destroy)( ThreadId tid, Bool isRT )
    if (VG_(clo_trace_signals))
       VG_(message)(
          Vg_DebugMsg, 
-         "VG_(signal_return) (thread %d): isRT=%d valid magic; RIP=%#llx\n",
+         "VG_(signal_return) (thread %u): isRT=%d valid magic; RIP=%#llx\n",
          tid, isRT, tst->arch.vex.guest_RIP);
 
    /* tell the tools */
