+++ coregrind/m_sigframe/sigframe-ppc64-linux.c
@@ -279,7 +279,7 @@ void VG_(sigframe_create)( ThreadId tid,
    if (0)
       VG_(printf)("pushed signal frame; %%R1 now = %#lx, "
                   "next %%CIA = %#llx, status=%d\n",
-		  sp, tst->arch.vex.guest_CIA, tst->status);
+		  sp, tst->arch.vex.guest_CIA, (Int)tst->status);
 }
 
 
@@ -345,7 +345,7 @@ void VG_(sigframe_destroy)( ThreadId tid, Bool isRT )
 
    if (VG_(clo_trace_signals))
       VG_(message)(Vg_DebugMsg,
-                   "vg_pop_signal_frame (thread %d): isRT=%d "
+                   "vg_pop_signal_frame (thread %u): isRT=%d "
                    "valid magic; EIP=%#llx\n",
                    tid, has_siginfo, tst->arch.vex.guest_CIA);
